Replace namespaces accordingly; Rename projects on main sln file.
