+++ samples/RenderDemo/App.xaml
@@ -2,6 +2,6 @@
   <Application.Styles>
     <StyleInclude Source="resm:Avalonia.Themes.Default.DefaultTheme.xaml?assembly=Avalonia.Themes.Default"/>
     <StyleInclude Source="resm:Avalonia.Themes.Default.Accents.BaseLight.xaml?assembly=Avalonia.Themes.Default"/>
-    <StyleInclude Source="resm:RenderTest.SideBar.xaml"/>
+    <StyleInclude Source="resm:RenderDemo.SideBar.xaml"/>
   </Application.Styles>
 </Application>
