+++ samples/RenderDemo/MainWindow.xaml
@@ -1,6 +1,6 @@
 <Window xmlns="https://github.com/avaloniaui"
         Title="AvaloniaUI Rendering Test"
-        xmlns:pages="clr-namespace:RenderTest.Pages"
+        xmlns:pages="clr-namespace:RenderDemo.Pages"
         Width="800"
         Height="600">
   <DockPanel>
